Minor Update
When dumping hashes with the drsuapi method it checks to see whether you have an existing file and asks for confirmation before overwriting. 

Added another option to stealth Mimikatz dump -hR AV which checks to see if Windows Defender Real Time Monitor is running. If it is it turns it off whilst running mimikatz and then turns it back on.